rename filename for fetch current price
