+++ apps/api/stratify-api/src/routes/assets/fetchCurrentPrice.ts
